feat: Adding presets
