+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 /lib
 *.tgz
+coverage
 
 # misc
 .DS_Store
