+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './formatToken'
+export * from './presets'
