+++ src/presets.ts
@@ -1,0 +1,41 @@
+interface PresetsType {
+  DATE_SHORT: string
+  DATE_MEDIUM: string
+  DATE_LONG: string
+  DATE_FULL: string
+  TIME: string
+  TIME_WITH_SECONDS: string
+  TIME_LONG: string
+  TIME_24: string
+  TIME_24_WITH_SECONDS: string
+  TIME_24_LONG: string
+  DATETIME_SHORT: string
+  DATETIME_SHORT_WITH_SECONDS: string
+  DATETIME_MEDIUM: string
+  DATETIME_MEDIUM_WITH_SECONDS: string
+  DATETIME_LONG: string
+  DATETIME_LONG_WITH_SECONDS: string
+  DATETIME_FULL: string
+  DATETIME_FULL_WITH_SECONDS: string
+}
+
+export const Presets: PresetsType = {
+  DATE_SHORT: 'date_short',
+  DATE_MEDIUM: 'date_med',
+  DATE_LONG: 'date_long',
+  DATE_FULL: 'date_full',
+  TIME: 'time',
+  TIME_WITH_SECONDS: 'time_with_seconds',
+  TIME_LONG: 'time_long',
+  TIME_24: 'time_24',
+  TIME_24_WITH_SECONDS: 'time_24_with_seconds',
+  TIME_24_LONG: 'time_24_long',
+  DATETIME_SHORT: 'datetime_short',
+  DATETIME_SHORT_WITH_SECONDS: 'datetime_short_with_seconds',
+  DATETIME_MEDIUM: 'datetime_medium',
+  DATETIME_MEDIUM_WITH_SECONDS: 'datetime_medium_with_seconds',
+  DATETIME_LONG: 'datetime_long',
+  DATETIME_LONG_WITH_SECONDS: 'datetime_long_with_seconds',
+  DATETIME_FULL: 'datetime_full',
+  DATETIME_FULL_WITH_SECONDS: 'datetime_full_with_seconds'
+}
