+++ tsconfig.browser.json
@@ -5,7 +5,7 @@
     "moduleResolution": "node",
     "declaration": true,
     "outDir": "./lib/esm/",
-    "lib": ["es2017"],
+    "lib": ["es2017", "dom"],
     "strict": true
   },
   "include": ["src"],
